backend/requirements.txt
matplotlib
flask
line-bot-sdk
python-dotenv
pyngrok
google-generativeai